突然死問題のみ未解決版
気圧センサ処理修正
加速度センサ処理修正
I2Cの処理を一つのファイルにまとめる
I2Cセンサ処理修正

+++ analyse_pressure.h
@@ -13,7 +13,7 @@
 #include    "system.h"
 
 extern void AnalysePressure_Init( void );
-extern void AnalysePressure_setNewRawPressure( float prs );
+extern void AnalysePressure_setNewRawPressure( int prs );
 extern bool AnalysePressure_catchEventOfPeriodic( uint8_t* midiValue );
 
 #if 0
